// TRABAJOS-COMPLEMENTARIOS/Parcial2-Taller3-Semana15/src/interfaces/INota.ts
// Generated by https://quicktype.io

export interface IResNota {
    sum:     number;
    notas: Nota[];
}

export interface Nota {
    _id?:    string;
    parcial: string;
    status?: boolean;
    nota: string;
    observacion:  string;
    estado: string;
    category?:  number;
    __v?:    number;
}